change location of storage directory
This commit removes the creation of a user and relies
on the btsync package to create a btsync system user and group.
Therefore the storage directory is moved to /var/lib/openmediavault-btsync
and is now also removed when the plugin is purged. This is so that
no old shared folders are still left when installing the plugin again.